refactor: move types to separate folder.
